refactor : getNotifications and Add countByReceiverId to NotificationRepository

+++ src/main/java/com/codeit/otboo/domain/notification/dto/NotificationDtoCursorResponse.java
@@ -13,7 +13,7 @@ public record NotificationDtoCursorResponse(
 	@NotNull
 	boolean hasNext,
 	@NotNull
-	int totalCount,
+	long totalCount,
 	String sortBy,
 	String sortDirection
 ) {
